# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClearLedger accounting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/reconciliations:
    get:
      operationId: listReconciliations
      tags: [reconciliations]
      summary: List reconciliations
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reconciliation"
    post:
      operationId: createReconciliation
      tags: [reconciliations]
      summary: Start a reconciliation
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReconciliationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reconciliation"
  /businesses/{businessId}/reconciliations/{reconciliationId}/run:
    post:
      operationId: runReconciliation
      tags: [reconciliations]
      summary: Run auto-reconciliation engine
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
        - in: path
          name: reconciliationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reconciliation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"
  /businesses/{businessId}/reconciliations/{reconciliationId}/complete:
    post:
      operationId: completeReconciliation
      tags: [reconciliations]
      summary: Mark reconciliation as complete
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
        - in: path
          name: reconciliationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reconciliation"

  # Reports
  /businesses/{businessId}/reports/profit-loss:
    get:
      operationId: getProfitLossReport
      tags: [reports]
      summary: Get P&L report
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
        - in: query
          name: startDate
          required: true
          schema:
            type: string
        - in: query
          name: endDate
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
  /businesses/{businessId}/reports/balance-sheet:
    get:
      operationId: getBalanceSheetReport
      tags: [reports]
      summary: Get balance sheet
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
        - in: query
          name: asOfDate
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceSheetReport"
  /businesses/{businessId}/reports/trial-balance:
    get:
      operationId: getTrialBalanceReport
      tags: [reports]
      summary: Get trial balance
      parameters:
        - in: path
          name: businessId
          required: true
          schema:
            type: integer
        - in: query
          name: asOfDate
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrialBalanceReport"

  # Dashboard
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        date:
          type: string
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [debit, credit]
        accountId:
          type: ["integer", "null"]
        accountName:
          type: ["string", "null"]
        source:
          type: string
          enum: [manual, plaid, upload]
        reconciled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - businessId
        - date
        - description
        - amount
        - type
        - source
        - reconciled
        - createdAt
...
    JournalLine:
      type: object
      properties:
        id:
          type: integer
        journalEntryId:
          type: integer
        accountId:
          type: integer
        accountName:
          type: ["string", "null"]
        debitAmount:
          type: number
        creditAmount:
          type: number
      required:
        - id
        - journalEntryId
        - accountId
        - debitAmount
        - creditAmount
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        date:
          type: string
        memo:
          type: ["string", "null"]
        createdAt:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/JournalLine"
      required:
        - id
        - businessId
        - date
        - createdAt
        - lines
...
    Reconciliation:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        bankAccountName:
          type: string
        statementDate:
          type: string
        openingBalance:
          type: number
        closingBalance:
          type: number
        status:
          type: string
          enum: [pending, in_progress, completed]
        createdAt:
          type: string
      required:
        - id
        - businessId
        - bankAccountName
        - statementDate
        - openingBalance
        - closingBalance
        - status
        - createdAt
    CreateReconciliationBody:
      type: object
      properties:
        bankAccountName:
          type: string
        statementDate:
          type: string
        openingBalance:
          type: number
        closingBalance:
          type: number
      required:
        - bankAccountName
        - statementDate
        - openingBalance
        - closingBalance
    MatchedPair:
      type: object
      properties:
        bankTransaction:
          $ref: "#/components/schemas/Transaction"
        journalEntry:
          $ref: "#/components/schemas/JournalEntry"
      required:
        - bankTransaction
        - journalEntry
    ReconciliationResult:
      type: object
      properties:
        matched:
          type: array
          items:
            $ref: "#/components/schemas/MatchedPair"
        unmatchedBank:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        unmatchedJournal:
          type: array
          items:
            $ref: "#/components/schemas/JournalEntry"
        summary:
          $ref: "#/components/schemas/ReconciliationSummary"
      required:
        - matched
        - unmatchedBank
        - unmatchedJournal
        - summary
    ReconciliationSummary:
      type: object
      properties:
        totalMatched:
          type: integer
        totalUnmatchedBank:
          type: integer
        totalUnmatchedJournal:
          type: integer
        openingBalance:
          type: number
        closingBalance:
          type: number
        difference:
          type: number
      required:
        - totalMatched
        - totalUnmatchedBank
        - totalUnmatchedJournal
        - openingBalance
        - closingBalance
        - difference

    # Report schemas
    ReportLineItem:
      type: object
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        accountCode:
          type: ["string", "null"]
        amount:
          type: number
      required:
        - accountId
        - accountName
        - amount
    ReportSection:
      type: object
      properties:
        title:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReportLineItem"
        total:
          type: number
      required:
        - title
        - items
        - total
    ProfitLossReport:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        income:
          $ref: "#/components/schemas/ReportSection"
        expenses:
          $ref: "#/components/schemas/ReportSection"
        netProfit:
          type: number
      required:
        - startDate
        - endDate
        - income
        - expenses
        - netProfit
    BalanceSheetReport:
      type: object
      properties:
        asOfDate:
          type: string
        assets:
          $ref: "#/components/schemas/ReportSection"
        liabilities:
          $ref: "#/components/schemas/ReportSection"
        equity:
          $ref: "#/components/schemas/ReportSection"
        totalLiabilitiesAndEquity:
          type: number
      required:
        - asOfDate
        - assets
        - liabilities
        - equity
        - totalLiabilitiesAndEquity
    TrialBalanceLineItem:
      type: object
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        accountCode:
          type: ["string", "null"]
        accountType:
          type: string
        debit:
          type: number
        credit:
          type: number
      required:
        - accountId
        - accountName
        - accountType
        - debit
        - credit
    TrialBalanceReport:
      type: object
      properties:
        asOfDate:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrialBalanceLineItem"
        totalDebits:
          type: number
        totalCredits:
          type: number
        isBalanced:
          type: boolean
      required:
        - asOfDate
        - items
        - totalDebits
        - totalCredits
        - isBalanced

    # Dashboard schemas
...