# lib/api-zod/src/generated/types/reportSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClearLedger accounting API
 * OpenAPI spec version: 0.1.0
 */
import type { ReportLineItem } from "./reportLineItem";

export interface ReportSection {
  title: string;
  items: ReportLineItem[];
  total: number;
}